com-nopcommerce-sw-4/src/main/java/com/nopcommerce/demo/pages/Login.java
package com.nopcommerce.demo.pages;

import com.nopcommerce.demo.utility.Utility;
import org.openqa.selenium.By;

public class Login extends Utility {
    By welcomeText = By.xpath("//h1[contains(text(),'Welcome, Please Sign In!')]");
    By emailField = By.id("Email");
    By passField = By.name("Password");
    By loginButton = By.xpath("//button[contains(text(),'Log in')]");
    By errorMessage = By.xpath("//body/div[6]/div[3]/div[1]/div[1]/div[1]/div[2]/div[1]/div[2]/form[1]/div[1]");


    public String getWelcomeText(){
        return getTextFromElement(welcomeText);
    }
    public void sendEmail(String emailF){
        sendTextToElement(emailField, emailF);
    }
    public void sendPass(String passF){
        sendTextToElement(passField, passF);
    }

    public void clickOnLoginButton(){
        clickOnElement(loginButton);
    }
    public String getErrorMessage(){
        return getTextFromElement(errorMessage);
    }
}
